# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Session revoked"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Service and database are reachable"
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/customers": {
      "get": {
        "operationId": "CustomersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerListDto"
                }
              }
            }
          }
        },
        "tags": [
          "customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        },
        "tags": [
          "customers"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "operationId": "CustomersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          },
          "404": {
            "description": "No such customer in this organization"
          }
        },
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          },
          "404": {
            "description": "No such customer in this organization"
          }
        },
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Customer archived"
          },
          "404": {
            "description": "No such customer in this organization"
          }
        },
        "tags": [
          "customers"
        ]
      }
    },
    "/api/measurement-definitions": {
      "get": {
        "operationId": "MeasurementDefinitionsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementDefinitionListDto"
                }
              }
            }
          }
        },
        "tags": [
          "measurement-definitions"
        ]
      }
    },
    "/api/customers/{customerId}/bikes": {
      "get": {
        "operationId": "CustomerBikesController_list",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeListDto"
                }
              }
            }
          },
          "404": {
            "description": "No such customer in this organization"
          }
        },
        "tags": [
          "bikes"
        ]
      },
      "post": {
        "operationId": "CustomerBikesController_create",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBikeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeDto"
                }
              }
            }
          },
          "404": {
            "description": "No such customer in this organization"
          }
        },
        "tags": [
          "bikes"
        ]
      }
    },
    "/api/bikes/{id}": {
      "get": {
        "operationId": "BikesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeDto"
                }
              }
            }
          },
          "404": {
            "description": "No such bike in this organization"
          }
        },
        "tags": [
          "bikes"
        ]
      },
      "patch": {
        "operationId": "BikesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBikeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeDto"
                }
              }
            }
          },
          "404": {
            "description": "No such bike in this organization"
          }
        },
        "tags": [
          "bikes"
        ]
      },
      "delete": {
        "operationId": "BikesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bike archived"
          },
          "404": {
            "description": "No such bike in this organization"
          }
        },
        "tags": [
          "bikes"
        ]
      }
    }
  },
  "info": {
    "title": "Fit My Bike API",
    "description": "Bike fitting studio management. All endpoints are scoped to the caller’s organization.",
    "version": "0.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fmb_session"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "FITTER"
            ]
          },
          "organization": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "slug"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "organization"
        ]
      },
      "CustomerListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "nullable": true
                },
                "phone": {
                  "type": "string",
                  "nullable": true
                },
                "dateOfBirth": {
                  "type": "string",
                  "nullable": true
                },
                "heightMm": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "nullable": true
                },
                "weightGrams": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "nullable": true
                },
                "notes": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "firstName",
                "lastName",
                "email",
                "phone",
                "dateOfBirth",
                "heightMm",
                "weightGrams",
                "notes",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "perPage": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "totalPages": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "nullable": true
          },
          "heightMm": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "weightGrams": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone",
          "dateOfBirth",
          "heightMm",
          "weightGrams",
          "notes",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "string",
                "maxLength": 255,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              }
            ],
            "nullable": true
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "nullable": true
          },
          "dateOfBirth": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              }
            ],
            "nullable": true
          },
          "heightMm": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "integer",
                "minimum": 1000,
                "maximum": 2500
              }
            ],
            "nullable": true
          },
          "weightGrams": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "integer",
                "minimum": 30000,
                "maximum": 250000
              }
            ],
            "nullable": true
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "nullable": true
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "string",
                "maxLength": 255,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              }
            ],
            "nullable": true
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "nullable": true
          },
          "dateOfBirth": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              }
            ],
            "nullable": true
          },
          "heightMm": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "integer",
                "minimum": 1000,
                "maximum": 2500
              }
            ],
            "nullable": true
          },
          "weightGrams": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  ""
                ]
              },
              {
                "type": "integer",
                "minimum": 30000,
                "maximum": 250000
              }
            ],
            "nullable": true
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "nullable": true
          }
        }
      },
      "MeasurementDefinitionListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "enum": [
                    "inseam",
                    "torso_length",
                    "arm_length",
                    "shoulder_width",
                    "foot_length",
                    "sternal_notch_height",
                    "forward_flexion",
                    "saddle_height",
                    "saddle_setback",
                    "saddle_angle",
                    "saddle_nose_to_bar",
                    "saddle_to_bar_drop",
                    "stem_length",
                    "stem_angle",
                    "spacer_stack",
                    "bar_width",
                    "bar_reach",
                    "crank_length",
                    "cleat_fore_aft"
                  ]
                },
                "label": {
                  "type": "string"
                },
                "category": {
                  "type": "string",
                  "enum": [
                    "BODY",
                    "BIKE"
                  ]
                },
                "unit": {
                  "type": "string",
                  "enum": [
                    "MM",
                    "DECIDEGREE",
                    "GRAM"
                  ]
                },
                "minValue": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "maxValue": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "helpText": {
                  "type": "string",
                  "nullable": true
                },
                "sortOrder": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "key",
                "label",
                "category",
                "unit",
                "minValue",
                "maxValue",
                "helpText",
                "sortOrder"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "BikeListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "customerId": {
                  "type": "string"
                },
                "brand": {
                  "type": "string",
                  "nullable": true
                },
                "model": {
                  "type": "string",
                  "nullable": true
                },
                "sizeLabel": {
                  "type": "string",
                  "nullable": true
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "ROAD",
                    "TT",
                    "GRAVEL",
                    "MTB",
                    "OTHER"
                  ]
                },
                "notes": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "customerId",
                "brand",
                "model",
                "sizeLabel",
                "type",
                "notes",
                "createdAt",
                "updatedAt"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateBikeDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "sizeLabel": {
            "type": "string",
            "maxLength": 40,
            "nullable": true
          },
          "type": {
            "default": "ROAD",
            "type": "string",
            "enum": [
              "ROAD",
              "TT",
              "GRAVEL",
              "MTB",
              "OTHER"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "nullable": true
          }
        }
      },
      "BikeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "sizeLabel": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "ROAD",
              "TT",
              "GRAVEL",
              "MTB",
              "OTHER"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "brand",
          "model",
          "sizeLabel",
          "type",
          "notes",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateBikeDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "sizeLabel": {
            "type": "string",
            "maxLength": 40,
            "nullable": true
          },
          "type": {
            "default": "ROAD",
            "type": "string",
            "enum": [
              "ROAD",
              "TT",
              "GRAVEL",
              "MTB",
              "OTHER"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "nullable": true
          }
        }
      }
    }
  }
}